Updated to build with stack/ghc 8.0
This updates the build to use Stack to build the components, and updates
the code to build on ghc 8.0 (dropping compatability with earlier versions;
or, at least, not ensuring backwards compatability).

The getproposals tool is being updated to work with the latest version of
the NASA/ADS API, but this work is not finished: that is, getproposals is
broken in this commit.